+++ tests/test_cmem_plugin_base.py
@@ -7,7 +7,6 @@ from cmem_plugin_base.dataintegration.entity import (
     EntityPath,
 )
 from cmem_plugin_base.dataintegration.description import Plugin, PluginParameter
-from cmem_plugin_base.dataintegration.discovery import discover_plugins, get_packages
 
 
 @Plugin(