+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [0.0.7] 2022-03-02
+
+### Added
+
+- parameter type validation and matching to internal types
+
 ## [0.0.6] 2022-02-28
 
 ### Added